// src/sdk/generated/nascar.d.ts
export interface components {
    schemas: {
        Unauthorized: {
            HttpStatusCode?: number;
            Code?: number;
            Description?: string;
            Help?: string;
        };
        Series: {
            /** @description The unique ID of this series */
            SeriesID?: number;
            /** @description The name of this series */
            Name?: string | null;
        };
        Race: {
            /** @description The unique ID of this race */
            RaceID?: number;
            /** @description The unique ID of the series that this race is associated with */
            SeriesID?: number;
            /** @description The name of the series that this race is associated with */
            SeriesName?: string | null;
            /** @description The season that this race is associated with */
            Season?: number;
            /** @description The name of this race */
            Name?: string | null;
            /** @description The day of this race */
            Day?: string | null;
            /** @description The date/time of this race */
            DateTime?: string | null;
            /** @description The name of the track for this race */
            Track?: string | null;
            /** @description The broadcast airing this race */
            Broadcast?: string | null;
            /** @description The unique ID of the driver who won this race */
            WinnerID?: number | null;
            /** @description The unique ID of the driver who won the qualifying pole */
            PoleWinnerID?: number | null;
            /** @description Indicates whether this race is currently in progress */
            IsInProgress?: boolean;
            /** @description Indicates whether this race is over */
            IsOver?: boolean;
            /** @description The date/time when this race was last updated */
            Updated?: string | null;
            /** @description The date/time when this race was created (in US Eastern Time) */
            Created?: string | null;
            /** @description The day that this is race is rescheduled for (if race was postponed) */
            RescheduledDay?: string | null;
            /** @description The date/time that this is race is rescheduled for (if race was postponed) */
            RescheduledDateTime?: string | null;
            /** @description Indicates whether this race has been canceled */
            Canceled?: boolean;
            /** @description Number of laps scheduled for this race */
            ScheduledLaps?: number | null;
            /** @description Total number of laps ran for this race */
            ActualLaps?: number | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}